- Safeguarding instance and plugin serialising - Moving mocking into submodule
