+++ CHANGES
@@ -6,6 +6,8 @@ Version 1.1.3
 
 - Adding Instance data to output
 - Adding Instance doc-string to output
+- Safeguarding Instance and Plugin serialising
+- Refactoring
 
 Version 1.1.2
 -------------
