+++ pyblish_endpoint/resource.py
@@ -18,7 +18,7 @@ import flask.ext.restful
 import flask.ext.restful.reqparse
 
 # Local library
-from service import current_service
+import service
 
 log = logging.getLogger("endpoint")
 
@@ -49,15 +49,40 @@ def format_error(error):
 
 
 def format_instance(instance):
+    """Serialise `instance`
+
+    For children to be visualised and modified,
+    they must provide an appropriate implementation
+    of __str__.
+
+    Data that isn't JSON compatible cannot be
+    visualised nor modified.
+
+    """
+
+    children = list()
+    for child in instance:
+        try:
+            child = str(child)
+        except:
+            child = "Invalid"
+        children.append(child)
+
+    data = dict()
+    for key, value in instance._data.iteritems():
+        try:
+            value = str(value)
+        except:
+            value = "Not supported"
+        data[key] = str(value)
+
     return {
         "name": instance.data("name"),
         "family": instance.data("family"),
         "objName": instance.name,
-        "category": getattr(instance, "category", None),
-        "nodes": [str(x) for x in instance],
-        "data": dict((str(k), str(v)) for k, v in instance._data.iteritems()),
+        "children": children,
+        "data": data,
         "publish": instance.data("publish"),
-        "doc": instance.data("doc")
     }
 
 
@@ -65,9 +90,9 @@ def format_plugin(plugin):
     formatted = {
         "name": plugin.__name__,
         "version": plugin.version,
+        "category": getattr(plugin, "category", None),
         "requires": plugin.requires,
         "order": plugin.order,
-        "active": False,
         "optional": plugin.optional,
         "doc": getattr(plugin, "doc", plugin.__doc__)
     }
@@ -114,8 +139,8 @@ class ApplicationApi(flask.ext.restful.Resource):
 
         data = {}
 
-        data.update(current_service().system())
-        data.update(current_service().versions())
+        data.update(service.current().system())
+        data.update(service.current().versions())
 
         return data, 200
 
@@ -174,7 +199,7 @@ class SessionApi(flask.ext.restful.Resource):
 
     def post(self):
         try:
-            status = current_service().init()
+            status = service.current().init()
         except Exception:
             _, _, exc_tb = sys.exc_info()
             tb = traceback.extract_tb(exc_tb)[-1]
@@ -195,7 +220,7 @@ class StateApi(flask.ext.restful.Resource):
     """
 
     def get(self):
-        return {"ok": True, "state": current_service().state}
+        return {"ok": True, "state": service.current().state}
 
     def post(self):
         parser = flask.ext.restful.reqparse.RequestParser()
@@ -214,7 +239,7 @@ class StateApi(flask.ext.restful.Resource):
             log.error(message)
             return {"ok": False, "message": message}, 500
 
-        current_service().state = state
+        service.current().state = state
         return {"ok": True, "state": state}, 200
 
 
@@ -226,7 +251,7 @@ class NextApi(flask.ext.restful.Resource):
     """
 
     def post(self):
-        result = current_service().next()
+        result = service.current().next()
 
         if result is not None:
             plugin = result["plugin"]
@@ -271,8 +296,8 @@ class PluginsListApi(flask.ext.restful.Resource):
 
         """
 
-        plugins = current_service().plugins
-        context = current_service().context
+        plugins = service.current().plugins
+        context = service.current().context
 
         families = set()
         for instance in context:
@@ -332,7 +357,7 @@ class InstancesListApi(flask.ext.restful.Resource):
 
         """
 
-        instances = list(current_service().context)
+        instances = list(service.current().context)
 
         response = []
         for instance in instances:
@@ -356,7 +381,7 @@ class InstancesApi(flask.ext.restful.Resource):
 
         """
 
-        instances = list(current_service().context)
+        instances = list(service.current().context)
 
         try:
             instance = find_instance(instance_id, instances)
@@ -383,7 +408,7 @@ class NodesListApi(flask.ext.restful.Resource):
 
         """
 
-        instances = list(current_service().context)
+        instances = list(service.current().context)
 
         try:
             instance = find_instance(instance_id, instances)
@@ -413,7 +438,7 @@ class DataListApi(flask.ext.restful.Resource):
 
         """
 
-        instances = list(current_service().context)
+        instances = list(service.current().context)
 
         try:
             instance = find_instance(instance_id, instances)
